% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_multivariate.R
\name{Alpha}
\alias{Alpha}
\title{Reliability analysis (Cronbach's \eqn{\alpha} and corrected item-total correlation)}
\usage{
Alpha(data, var, items, vars = NULL, rev = NULL)
}
\description{
Reliability analysis (Cronbach's \eqn{\alpha} and corrected item-total correlation)
}
