% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_multivariate.R
\name{CFA}
\alias{CFA}
\title{Confirmatory factor analysis (CFA)}
\usage{
CFA(data, model = "A =~ a[1:5]; B =~ b[c(1,3,5)]; C =~ c1 + c2 + c3",
  highorder = "", orthogonal = FALSE, missing = "listwise",
  style = c("jmv", "lavaan"), CI = FALSE, MI = FALSE, plot = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table}.}

\item{model}{Model formula.}

\item{highorder}{Optional. High-order factor. Default is \code{""}.}
}
\description{
Confirmatory factor analysis (CFA)
}
\examples{
data.cfa=lavaan::HolzingerSwineford1939
CFA(data.cfa, "Visual =~ x[1:3]; Textual =~ x[c(4,5,6)]; Speed =~ x7 + x8 + x9", plot=T)
CFA(data.cfa, model="
    Visual =~ x[1:3]
    Textual =~ x[c(4,5,6)]
    Speed =~ x7 + x8 + x9
    ", highorder="Ability", plot=T)

data.bfi=psych::bfi
data.bfi=data.bfi[complete.cases(data.bfi),]
CFA(data.bfi, "E =~ E[1:5]; A =~ A[1:5]; C =~ C[1:5]; N =~ N[1:5]; O =~ O[1:5]", plot=T)
}
