% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{CI}
\alias{CI}
\title{Compute 95\% confidence interval (CI) of a variable}
\usage{
CI(var, nsmall = 2, empirical = TRUE)
}
\arguments{
\item{var}{Variable (e.g., \code{data$var}) or numeric vector.}

\item{nsmall}{Number of decimal places of output. Default is 3.}

\item{empirical}{\code{TRUE} (default, return 2.5\% and 97.5\% percentiles) or \code{FALSE} (basd on \emph{t} distribution).
If \code{TRUE}, directly extract CI from data (instead of computing CI based on normal or \emph{t} distribution).}
}
\description{
Compute 95\% confidence interval (CI) of a variable
}
\examples{
CI(bfi$age)
CI(bfi$age, emp=F)  # NOTE: "emp" can partially match "empirical"
}
