% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR-stats_1_basic.R
\name{cor_multilevel}
\alias{cor_multilevel}
\title{Multilevel correlations (within-level and between-level).}
\usage{
cor_multilevel(data, group, digits = 3)
}
\arguments{
\item{data}{Data frame.}

\item{group}{Grouping variable.}

\item{digits}{Number of decimal places of output. Defaults to \code{3}.}
}
\value{
Invisibly return a list of results.
}
\description{
Multilevel correlations (within-level and between-level).
For details, see description in \code{\link{HLM_ICC_rWG}}.
}
\examples{
# see https://psychbruce.github.io/supp/CEM

}
\seealso{
\code{\link{Corr}}

\code{\link{HLM_ICC_rWG}}
}
