% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{Corr_diff}
\alias{Corr_diff}
\title{Test the difference between two correlations (independent / nonindependent)}
\usage{
Corr_diff(r1, n1, r2, n2, rcov = NULL)
}
\arguments{
\item{r1, r2}{Correlation coefficients (Pearson's \emph{r}).}

\item{n1, n2}{Sample sizes.}

\item{rcov}{[Optional] Only for nonindependent \emph{r}s:

\code{r1} is r(X,Y),

\code{r2} is r(X,Z),

then, as Y and Z are also correlated,

we should also consider \code{rcov}: r(Y,Z)}
}
\value{
Invisibly return the \emph{p} value.
}
\description{
Test the difference between two correlations (independent / nonindependent)
}
\examples{
# two independent rs (X~Y vs. Z~W)
Corr_diff(0.20, 100, 0.45, 100)

# two nonindependent rs (X~Y vs. X~Z, with Y and Z also correlated [rcov])
Corr_diff(0.20, 100, 0.45, 100, rcov=0.80)
}
