% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_plot.R
\name{drawChinaMap}
\alias{drawChinaMap}
\title{Draw standard China maps}
\usage{
drawChinaMap(provdata = NULL, citydata = NULL, var = NA,
  multiply = 1, nsmall = 0, colors = "Blues", direc = 1,
  addlabel = "", labelseg = "\\n", tag = "", title = var,
  guidetitle = var, bordersize = 0.2, bordercolor = "grey70",
  na.color = "grey90", filename = "ChinaMap.png", dpi = 500)
}
\description{
Draw standard China maps
}
\examples{
drawChinaMap() # draw a demo map
drawChinaMap(provdata_demo, var="geoE", nsmall=1, filename="ChinaMap1.png")
drawChinaMap(provdata_demo, var="geoN", nsmall=1, colors="Reds", direc=-1, filename="ChinaMap2.png")
}
