% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_utils.R
\name{dtime}
\alias{dtime}
\title{Timer (compute time difference).}
\usage{
dtime(t0, unit = "secs", digits = 0)
}
\arguments{
\item{t0}{Time at the beginning.}

\item{unit}{Options: \code{"auto"}, \code{"secs"}, \code{"mins"}, \code{"hours"}, \code{"days"}, \code{"weeks"}. Defaults to \code{"secs"}.}

\item{digits}{Number of decimal places of output. Defaults to \code{0}.}
}
\value{
A character string of time difference.
}
\description{
Timer (compute time difference).
}
\examples{
\dontrun{

t0 = Sys.time()
dtime(t0)
}

}
