% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_multivariate.R
\name{EFA}
\alias{EFA}
\title{Exploratory factor analysis (EFA)}
\usage{
EFA(data, var, items, vars = NULL, method = "parallel",
  rotation = "varimax", fixedFactors = 1, hideBelow = 0.4)
}
\description{
Exploratory factor analysis (EFA)
}
