% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_manova.R
\name{eta_sq_ci}
\alias{eta_sq_ci}
\title{Compute the confidence interval (CI) of partial eta^2 in ANOVA}
\usage{
eta_sq_ci(F.value, df.1, df.2, conf.level = 0.9, return = "all")
}
\description{
Compute the confidence interval (CI) of partial eta^2 in ANOVA
}
