% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_stats_04_regress.R
\name{formula_expand}
\alias{formula_expand}
\title{Expand all interaction terms in a formula.}
\usage{
formula_expand(formula)
}
\arguments{
\item{formula}{R formula.}
}
\description{
Expand all interaction terms in a formula.
}
\examples{
formula_expand(y ~ a*b*c)

}
