% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR-stats_4_regress.R
\name{formula_paste}
\alias{formula_paste}
\title{Paste a formula into a string.}
\usage{
formula_paste(formula)
}
\arguments{
\item{formula}{R formula.}
}
\value{
A character string indicating the formula.
}
\description{
Paste a formula into a string.
}
\examples{
formula_paste(y ~ x)
formula_paste(y ~ x + (1 | g))

}
