% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{Freq}
\alias{Freq}
\title{Frequency statistics with histogram and density plot}
\usage{
Freq(var, label = NULL, sort = "", nsmall = 1)
}
\arguments{
\item{var}{Vector or variable.}

\item{label}{[optional] A vector re-defining the labels of values.}

\item{sort}{\code{""} (default, sorted by raw order), \code{"-"} (decreasing order), or \code{"+"} (increasing order).}

\item{nsmall}{Number of decimal places of output. Default is 1.}
}
\description{
Frequency statistics with histogram and density plot
}
\examples{
Freq(bfi$education)
Freq(bfi$gender, label=c("Male", "Female"))
Freq(bfi$age)
}
