% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_regress.R
\name{grand_mean_center}
\alias{grand_mean_center}
\title{Grand-mean centering}
\usage{
grand_mean_center(data, vars)
}
\arguments{
\item{data}{a \code{data.frame} or \code{data.table}}

\item{vars}{a character vector specifying which variable(s) to be centered}
}
\value{
A \code{data.frame} or \code{data.table}, with the same type as the input data.
}
\description{
Compute grand-mean centered variables. Usually used for HLM level-2 variables.
}
\seealso{
\code{\link{group_mean_center}}
}
