% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_stats_05_advance.R
\name{granger_causality}
\alias{granger_causality}
\title{Granger causality test (multivariate).}
\usage{
granger_causality(varmodel, var.y = NULL, var.x = NULL, check.dropped = FALSE)
}
\arguments{
\item{varmodel}{VAR model fitted using the \code{\link[vars:VAR]{vars::VAR()}} function.}

\item{var.y, var.x}{[optional] Default is \code{NULL} (all variables).
If specified, then perform tests for specific variables.
Values can be a single variable (e.g., \code{"X"}),
a vector of variables (e.g., \code{c("X1", "X2")}),
or a string containing regular expression (e.g., \code{"X1|X2"}).}

\item{check.dropped}{Check dropped variables. Default is \code{FALSE}.}
}
\value{
A data frame of results.
}
\description{
Granger test of predictive causality (between multivariate time series)
based on vector autoregression (\code{\link[vars:VAR]{VAR}}) model.
Its output resembles the output of the \code{vargranger}
command in Stata (but here using an \emph{F} test).
}
\details{
The Granger causality test (based on VAR model) examines whether
the lagged values of a predictor (or predictors)
have any incremental role in predicting an outcome
if controlling for the lagged values of the outcome itself.
It compares the two models by using an \emph{F} test.
}
\seealso{
\code{\link{ccf_plot}},
\code{\link{granger_test}}
}
