% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_stats_05_advance.R
\name{granger_test}
\alias{granger_test}
\title{Granger causality test.}
\usage{
granger_test(
  formula,
  data,
  lags = 1:5,
  test.reverse = FALSE,
  na.action = na.omit
)
}
\arguments{
\item{formula}{Model formula like \code{y ~ x}.}

\item{data}{Data frame.}

\item{lags}{Time lags. Default is \code{1:5}.}

\item{test.reverse}{Whether to test reverse causality. Default is \code{FALSE}.}

\item{na.action}{What to do with missing values. Default is \code{na.omit}.}
}
\description{
Granger test of predictive causality (between two time-series variables)
using the \code{lmtest::\link[lmtest]{grangertest}} function.
}
\details{
The Granger causality test can examine whether
the lagged values of a predictor
have any incremental role in predicting an outcome
if controlling for
the lagged values of the outcome itself.
It compares the two models by using an \emph{F} test.
}
\examples{
(d=lmtest::ChickEgg)  # time-series data
granger_test(chicken ~ egg, data=d)
granger_test(chicken ~ egg, data=d, lags=1:10, test.reverse=TRUE)

}
\seealso{
\code{\link{ccf_plot}}
}
