% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{\%allin\%}
\alias{\%allin\%}
\title{A simple extension of \code{\%in\%}.}
\usage{
x \%allin\% vector
}
\arguments{
\item{x}{Numeric or character vector.}

\item{vector}{Numeric or character vector.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
A simple extension of \code{\%in\%}.
}
\examples{
1:2 \%allin\% 1:3  # TRUE
3:4 \%allin\% 1:3  # FALSE

}
\seealso{
\code{\link[base:match]{\%in\%}},
\code{\link{\%anyin\%}},
\code{\link{\%nonein\%}},
\code{\link{\%partin\%}}
}
