% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{\%anyin\%}
\alias{\%anyin\%}
\title{A simple extension of \code{\%in\%}}
\usage{
x \%anyin\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link{Long vectors} are not supported.}
}
\description{
A simple extension of \code{\%in\%}
}
\seealso{
\code{\link[base]{match}} (\code{\%in\%})
}
