% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{\%nonein\%}
\alias{\%nonein\%}
\title{A simple extension of \code{\%in\%}.}
\usage{
x \%nonein\% vector
}
\arguments{
\item{x}{Numeric or character vector.}

\item{vector}{Numeric or character vector.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
A simple extension of \code{\%in\%}.
}
\examples{
3:4 \%nonein\% 1:3  # FALSE
4:5 \%nonein\% 1:3  # TRUE

}
\seealso{
\code{\link[base]{match}} (\code{\%in\%}), \code{\link{\%allin\%}}, \code{\link{\%anyin\%}}, \code{\link{\%partin\%}}
}
