% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{\%partin\%}
\alias{\%partin\%}
\title{A simple extension of \code{\%in\%}}
\usage{
pattern \%partin\% vector
}
\arguments{
\item{pattern}{a character string containing \strong{regular expressions} to be matched in the given character vector.}

\item{vector}{a character vector.}
}
\description{
A simple extension of \code{\%in\%}
}
\seealso{
\code{\link[base]{match}} (\code{\%in\%})
}
