% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{\%^\%}
\alias{\%^\%}
\title{Paste strings together.}
\usage{
x \%^\% y
}
\arguments{
\item{x, y}{Any objects, usually a numeric or character string or vector.}
}
\value{
A character string/vector of the pasted values.
}
\description{
Paste strings together. A wrapper of \code{paste0()}.
Why \code{\%^\%}? Because typing \code{\%} and \code{^} is pretty easy by
pressing \strong{Shift + 5 + 6 + 5}.
}
\examples{
"He" \%^\% "llo"
"X" \%^\% 1:10
"Q" \%^\% 1:5 \%^\% letters[1:5]

}
