% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_regress.R
\name{group_mean_center}
\alias{group_mean_center}
\title{Group-mean centering}
\usage{
group_mean_center(data, vars, by)
}
\arguments{
\item{data}{\code{data.frame} or \code{data.table}.}

\item{vars}{Variable(s) to be centered.}

\item{by}{Grouping variable.}
}
\description{
Compute group-mean centered variables. Usually used for HLM level-1 variables.
}
\seealso{
\code{\link{grand_mean_center}}
}
