% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_regress.R
\name{HLMs_formulas}
\alias{HLMs_formulas}
\title{Generate all possible combinations of random effects}
\usage{
HLMs_formulas(formula.full.model)
}
\description{
Generate all possible combinations of random effects
}
\examples{
f = y ~ a * b * c + (a * b * c | sub) + (a * b * c | item)
fs=HLMs_formulas(f)
}
