% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_regress.R
\name{HLMs_run}
\alias{HLMs_run}
\title{Run many HLMs}
\usage{
HLMs_run(formulas.text, data, family = NULL)
}
\description{
Run many HLMs
}
\examples{
mf=HLMs_run(HLMs_formulas(Reaction ~ Days + (Days | Subject)), data=sleepstudy)
}
