% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_regress.R
\name{HLMs_run_parallel}
\alias{HLMs_run_parallel}
\title{Run many HLMs (parallel version, much faster than \code{HLMs_run()})}
\usage{
HLMs_run_parallel(formulas.text, data, family = NULL, cores = 4)
}
\value{
A data.table ordered by \code{singular}, \code{BIC}, and \code{AIC}.
}
\description{
Run many HLMs (parallel version, much faster than \code{HLMs_run()})
}
\note{
Collaborated with \href{https://github.com/usplos}{Guang-Yao Zhang}
}
\seealso{
\code{\link{HLMs_run}}
}
