% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_stats_05_advance.R
\name{lavaan_summary}
\alias{lavaan_summary}
\title{Tidy report of lavaan model.}
\usage{
lavaan_summary(
  lavaan,
  ci = c("raw", "boot", "bc.boot", "bca.boot"),
  nsim = 100,
  seed = 1
)
}
\description{
Tidy report of lavaan model.
}
