% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_multivariate.R
\name{LOOKUP}
\alias{LOOKUP}
\title{Search, match, and look up values}
\usage{
LOOKUP(data, vars, data.ref, vars.ref, vars.lookup)
}
\arguments{
\item{data}{a data.frame or data.table.}

\item{vars}{a character or character vector, specifying the variable(s) to be searched in \code{data}.}

\item{data.ref}{a reference data containing both the reference variable(s) and the lookup variable(s).}

\item{vars.ref}{a character or character vector (with the \strong{same length and order} as \code{vars}),
specifying the reference variable(s) to be matched in \code{data.ref}.}

\item{vars.lookup}{a character or character vector, specifying the variable(s) to be looked up and returned from \code{data.ref}.}
}
\value{
A new data.frame or data.table, with the lookup values added.
If multiple values were simultaneously matched, a warning message would be printed.
Then you may check if there was anything wrong in \code{data.ref}, and/or re-define \code{vars} and \code{vars.ref}.
}
\description{
Search, match, and look up values
}
\examples{
dict=data.table(City=rep(c("A", "B", "C"), each=5),
                Year=rep(2013:2017, times=3),
                GDP=RANDBETWEEN(1000:2000, n=15, seed=1),
                PM2.5=RANDBETWEEN(10:300, n=15, seed=1))
dict

data=data.table(sub=1:5,
                city=c("A", "A", "B", "C", "C"),
                year=c(2013, 2014, 2015, 2016, 2017))
data

LOOKUP(data, c("city", "year"), dict, c("City", "Year"), "GDP")
LOOKUP(data, c("city", "year"), dict, c("City", "Year"), c("GDP", "PM2.5"))
LOOKUP(data, "city", dict, "City", "GDP")  # return with a warning
}
