% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_manova.R
\name{MANOVA}
\alias{MANOVA}
\title{Multivariate ANOVA}
\usage{
MANOVA(data, id = NULL, dv = NULL, dvs = NULL, dvs.pattern = "",
  between = NULL, within = NULL, covariate = NULL,
  factorize = ifelse(is.null(covariate), TRUE, FALSE),
  correction = "none")
}
\arguments{
\item{correction}{Sphericity correction method to adjust the degrees of freedom (\emph{df}) when the sphericity assumption is violated. Default is \code{"none"}.
If Mauchly's test of sphericity is significant, you may set it to \code{"GG"} (Greenhouse-Geisser) or \code{"HF"} (Huynh-Feldt).}
}
\description{
Easily perform MANOVA (between-subject, within-subject, and mixed design).
}
\details{
\strong{Demo datasets:}
\describe{
  \item{1. Between-subject design}{
    \itemize{
      \item \code{between.1} - 1
      \item \code{between.2} - 2
      \item \code{between.3} - 3
    }
  }
  \item{2. Within-subject design}{
    \itemize{
      \item \code{within.1} - 1
      \item \code{within.2} - 2
      \item \code{within.3} - 3
    }
  }
  \item{3. Mixed design}{
    \itemize{
      \item \code{mixed.2_1b1w} - 1
      \item \code{mixed.3_1b2w} - 2
      \item \code{mixed.3_2b1w} - 3
    }
  }
}
}
\examples{
## Between-Subject Design

View(between.1)
MANOVA(data=between.1, dv="SCORE", between="A")

View(between.2)
MANOVA(data=between.2, dv="SCORE", between=c("A", "B"))

View(between.3)
MANOVA(data=between.3, dv="SCORE", between=c("A", "B", "C"))


## Within-Subject Design

View(within.1)
MANOVA(data=within.1, dvs="A1:A4", dvs.pattern="A(.)",
       within="A")

View(within.2)
MANOVA(data=within.2, dvs="A1B1:A2B3", dvs.pattern="A(.)B(.)",
       within=c("A", "B"))

View(within.3)
MANOVA(data=within.3, dvs="A1B1C1:A2B2C2", dvs.pattern="A(.)B(.)C(.)",
       within=c("A", "B", "C"))


## Mixed Design

View(mixed.2_1b1w)
MANOVA(data=mixed.2_1b1w, dvs="B1:B3", dvs.pattern="B(.)",
       between="A", within="B")
MANOVA(data=mixed.2_1b1w, dvs="B1:B3", dvs.pattern="B(.)",
       between="A", within="B", correction="GG")
}
\seealso{
\code{\link{EMMEANS}}
}
