% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_regress.R
\name{med}
\alias{med}
\title{Mediation analysis based on regression coefficients and standard errors}
\usage{
med(a, SEa, b, SEb, direct = NULL, total = NULL, cov_ab = 0,
  seed = 1, rep = 50000, nsmall = 3)
}
\arguments{
\item{a}{Path \strong{a} (X -> Mediator).}

\item{SEa}{\emph{SE} of path \strong{a}.}

\item{b}{Path \strong{b} (Mediator -> Y).}

\item{SEb}{\emph{SE} of path \strong{b}.}

\item{direct}{\strong{[optional]} Path \strong{c'} (X -> Y \strong{direct} effect, with M also included in model).}

\item{total}{\strong{[optional]} Path \strong{c} (X -> Y \strong{total} effect, without M).}

\item{cov_ab}{Covariance between \strong{a} and \strong{b}.

\href{http://www.quantpsy.org/medmc/medmc.htm}{Selig & Preacher (2008)}:

If you use SEM, path analysis, multilevel modeling, or some other multivariate method to obtain both a and b from a single model, then cov(a,b) can be found in the asymptotic covariance matrix of the parameter estimates.
If you use regression to obtain a and b in separate steps, then cov(a,b) = 0.}

\item{seed}{Random seed. Default is 1.}

\item{rep}{Number of repetitions for Monte Carlo simulation. Default is 50,000. More than 1,000 are recommended.}

\item{nsmall}{Number of decimal places of output. Default is 3.}
}
\description{
Estimating indirect effect from regression coefficients and standard errors (\emph{SE}) by using Sobel test and Monte Carlo simulation.

Total effect (\strong{c}) = Direct effect (\strong{c'}) + Indirect effect (\strong{a*b})
}
\examples{
med(a=1.50, SEa=0.50, b=2.00, SEb=0.80)
med(a=1.50, SEa=0.50, b=2.00, SEb=0.80, total=4.50)
med(a=1.50, SEa=0.50, b=2.00, SEb=0.80, direct=1.50)
}
\references{
Sobel, M. E. (1982). Asymptotic confidence intervals for indirect effects in Structural Equation Models. \emph{Sociological Methodology, 13,} 290-312.

Selig, J. P., & Preacher, K. J. (2008). Monte Carlo method for assessing mediation: An interactive tool for creating confidence intervals for indirect effects. \url{http://www.quantpsy.org/medmc/medmc.htm}
}
