% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_regress.R
\name{med_summary}
\alias{med_summary}
\title{Better report of mediation analysis based on the \code{mediation} package}
\usage{
med_summary(model, nsmall = 3)
}
\arguments{
\item{model}{Mediation model built with the \code{mediation} package (see \code{\link[mediation]{mediate}}).}

\item{nsmall}{Number of decimal places of output. Default is 3.}
}
\description{
Better report of mediation analysis based on the \code{\link[mediation]{mediation}} package.
}
