% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_regress.R
\name{model_check}
\alias{model_check}
\title{Check many assumptions for (OLS and multilevel) regression models}
\usage{
model_check(model, plot = TRUE)
}
\arguments{
\item{model}{A model object (fitted by \code{lm, glm, lmer, glmer, ...}).}

\item{plot}{Visualize the check results. Default is \code{TRUE}.}
}
\description{
Based on the functions in \code{performance} (see \code{performance::\link[performance]{check_model}}), it checks for
1) multivariate normality,
2) multicollinearity (VIF),
3) homoscedasticity (vs. heteroscedasticity),
4) independence of residuals (vs. autocorrelation).
}
\examples{
lm=lm(Temp ~ Month + Day + Wind + Solar.R, data=airquality)
model_check(lm)

library(lmerTest)
hlm.2=lmer(Preference ~ Sweetness + Gender * Age + Frequency + (Sweetness | Consumer) + (1 | Product), data=carrots)
model_check(hlm.2)
}
