% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_stats_01_basic.R
\name{p.trans}
\alias{p.trans}
\title{Transform p values.}
\usage{
p.trans(p, nsmall.p = 3)
}
\arguments{
\item{p}{\emph{p} value.}

\item{nsmall.p}{Number of decimal places of \emph{p} value. Default is \code{3}.}
}
\description{
Transform p values.
}
\examples{
p.trans(c(1, 0.1, 0.05, 0.01, 0.001, 0.0001, 0.0000001, 1e-50)) \%>\% data.table(p=.)

}
\seealso{
\code{\link{p.trans2}}
}
