% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{pkg_depend}
\alias{pkg_depend}
\title{Check dependencies of a package}
\usage{
pkg_depend(pkg)
}
\arguments{
\item{pkg}{Package name.}
}
\description{
Check dependencies of a package
}
\examples{
pkg.depend("sjstats")
pkg.depend("MuMIn")
}
