% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{print_table}
\alias{print_table}
\title{Print three-line table}
\usage{
print_table(x, row.names = TRUE, nsmalls = 3)
}
\arguments{
\item{x}{Matrix, data.frame, data.table, or any model object (e.g., \code{lm, glm, lmer, glmer, ...}).}

\item{row.names}{\code{TRUE} (default) or \code{FALSE}. Whether to print row names.}

\item{nsmalls}{A number or numeric vector specifying the number of decimal places of output. Default is 3.}
}
\description{
Print three-line table
}
\examples{
model=lm(Temp ~ Month + Day + Wind + Solar.R, data=airquality)
print_table(model)
}
