% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{RANDBETWEEN}
\alias{RANDBETWEEN}
\title{Randomly sampling (like Excel's function \code{RANDBETWEEN})}
\usage{
RANDBETWEEN(range, n = 1, seed = NULL)
}
\arguments{
\item{range}{numeric vector, or character vector.}

\item{n}{sample size.}

\item{seed}{random seed.}
}
\description{
Randomly sampling (like Excel's function \code{RANDBETWEEN})
}
\examples{
RANDBETWEEN(1:10, n=1000000) \%>\% Freq()
RANDBETWEEN(LETTERS, n=1000000) \%>\% Freq()
}
