% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{RGB}
\alias{RGB}
\title{A simple extension of \code{rgb()}}
\usage{
RGB(r, g, b, alpha)
}
\arguments{
\item{r, g, b}{an integar \emph{[0, 255]} indicating red, green, and blue}

\item{alpha}{a value \emph{[0, 1]} indicating color opacity (transparency); if not specified, an opaque colour will be generated}
}
\value{
\code{"#rrggbb"} or \code{"#rrggbbaa"}
}
\description{
A simple extension of \code{rgb()}
}
\examples{
RGB(255, 0, 0)  # red: "#FF0000"
RGB(255, 0, 0, )  # red with 80\\\% opacity: "#FF0000CC"
}
