% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{Run}
\alias{Run}
\title{Run code parsed from text.}
\usage{
Run(text)
}
\arguments{
\item{text}{Character strings for running.
You can use \code{"{ }"} to insert any R object in the environment.}
}
\value{
No return value.
}
\description{
Run code parsed from text.
}
\examples{
Run("a=1")
a

Run("a={a+1}")
a

}
