% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_multivariate.R
\name{scaler}
\alias{scaler}
\title{Min-max scaling (min-max normalization)}
\usage{
scaler(v, min = 0, max = 1)
}
\arguments{
\item{v}{Variable (numeric vector).}

\item{min}{Minimum value (default is 0).}

\item{max}{Maximum value (default is 1).}
}
\description{
This function resembles \code{\link[bruceR]{RESCALE}},
it just equals to \code{RESCALE(var, to=0:1)}.
}
\examples{
scaler(1:5)
# the same: RESCALE(1:5, to=0:1)
}
