% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR-stats_6_plot.R
\name{show_colors}
\alias{show_colors}
\title{Show colors.}
\usage{
show_colors(colors)
}
\arguments{
\item{colors}{Color names.

e.g.,
\itemize{
  \item \code{"red"} (R base color names)
  \item \code{"#FF0000"} (hex color names)
  \item \code{see::social_colors()}
  \item \code{viridis::viridis_pal()(10)}
  \item \code{RColorBrewer::brewer.pal(name="Set1", n=9)}
  \item \code{RColorBrewer::brewer.pal(name="Set2", n=8)}
  \item \code{RColorBrewer::brewer.pal(name="Spectral", n=11)}
}}
}
\value{
A \code{gg} object.
}
\description{
Show colors.
}
\examples{
show_colors("blue")
show_colors("#0000FF")  # blue (hex name)
show_colors(RGB(0, 0, 255))  # blue (RGB)
show_colors(see::social_colors())
show_colors(see::pizza_colors())

}
