% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_stats_01_basic.R
\name{sig.trans}
\alias{sig.trans}
\title{Transform \emph{p} value to significance code.}
\usage{
sig.trans(p)
}
\arguments{
\item{p}{\emph{p} value.}
}
\value{
A character string of significance code.
}
\description{
Transform \emph{p} value to significance code.
}
\examples{
sig.trans(c(1, 0.09, 0.049, 0.009, 0.001, 0.0001, 1e-50)) \%>\% data.table(sig=.)

}
