% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_regress.R
\name{simple_slope}
\alias{simple_slope}
\title{Simple-slope analysis based on \emph{b} and \emph{SE}}
\usage{
simple_slope(b, SEb, bmod, SDmod, df, nsmall = 3)
}
\arguments{
\item{b}{Coefficient of X (main predictor).}

\item{SEb}{\emph{SE} of b.}

\item{bmod}{Coefficient of moderator.}

\item{SDmod}{\emph{SD} of moderator (not \emph{SE}), used for calculating simple slopes of X with moderator at "\emph{M} + \emph{SD}" and "\emph{M} - \emph{SD}".}

\item{df}{Residual degree of freedom.}

\item{nsmall}{Number of decimal places of output. Default is 3.}
}
\description{
\strong{WARNING}: This function is NOT optimal.
I suggest using the \code{interactions} package, see \code{interactions::\link[interactions]{sim_slopes}}
}
\examples{
simple_slope(b=1.5, SEb=0.5, bmod=0.9, SDmod=1.0, df=300)
}
