% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\docType{package}
\name{bruceR-package}
\alias{bruceR-package}
\alias{bruceR}
\title{bruceR: BRoadly Useful Collections and Extensions of R functions}
\description{
Packing many useful R packages and functions into one package.
Making data analyses and statistics more elegant.
}
\note{
\emph{Please always use \href{https://www.rstudio.com/products/rstudio/download/preview/}{RStudio} as an \href{https://en.wikipedia.org/wiki/Integrated_development_environment}{IDE} instead of using the raw R.}
}
\section{Main Functions in \code{bruceR}}{

\code{bruceR} includes functions for 1) basic analyses, 2) multivariate computing (e.g., scale mean),
3) questionnaire analyses (e.g., reliability, EFA, CFA), 4) advanced outputs for LM, GLM, HLM (LMM), and GLMM,
5) nice themes for \code{ggplot2}, and 6) drawing standard China maps.

\describe{
  \item{\strong{Basic Functions}}{
      \code{\link{set.wd}}

      \code{\link{set.seeds}}

      \code{\link{Print}} (see also \code{\link{Glue}})

      \code{\link{print_table}}

      \code{\link{p}}

      \code{\link{Describe}} (descriptive statistics)

      \code{\link{Freq}} (frequency statistics with plot)

      \code{\link{Corr}} (correlation analysis with plot)
  }
  \item{\strong{Multivariate Computing}}{
      \code{\link{COUNT}}

      \code{\link{SUM}}

      \code{\link{MEAN}}

      \code{\link{STD}}

      \code{\link{CONSEC}}

      \code{\link{RECODE}}

      \code{\link{RESCALE}}

      \code{\link{LOOKUP}}
  }
  \item{\strong{Reliability, EFA, & CFA}}{
      \code{\link{Alpha}}

      \code{\link{EFA}}

      \code{\link{CFA}}
  }
  \item{\strong{MANOVA, Simple Effects, & Multiple Comparisons}}{
      \emph{(Coming soon...)}
  }
  \item{\strong{Linear Models}}{
      \code{\link{regress}} (do many kinds of regression analyses in one function)

      \code{\link{grand_mean_center}}

      \code{\link{group_mean_center}}

      \code{\link{GLM_summary}}

      \code{\link{HLM_summary}}
  }
  \item{\strong{Mediation & Moderation}}{
      \code{\link{med}}

      \code{\link{mod_med}}

      \code{\link{simple_slope}}
  }
  \item{\strong{Plotting with \code{ggplot2}}}{
      \code{\link{theme_bruce}}

      \code{\link{drawChinaMap}}
  }
}
}

\author{
\href{https://www.zhihu.com/people/psychbruce/}{Han-Wu-Shuang (Bruce) Bao} <E-mail: \email{baohws@psych.ac.cn} or \email{psychbruce@qq.com}>
}
