% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{CI}
\alias{CI}
\title{Compute 95\% confidence interval (CI) of a variable}
\usage{
CI(var, nsmall = 2)
}
\arguments{
\item{var}{a variable (e.g., data$var) or a numeric vector}
}
\description{
Compute 95\% confidence interval (CI) of a variable
}
