% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_stats_01_basic.R
\name{Corr}
\alias{Corr}
\title{Correlation analysis with test and plot.}
\usage{
Corr(
  data,
  method = "pearson",
  nsmall = 2,
  p.adjust = "none",
  plot = TRUE,
  plot.range = c(-1, 1),
  plot.color = c("#B52127", "white", "#2171B5"),
  save.file = NULL,
  width = 8,
  height = 6,
  dpi = 500
)
}
\arguments{
\item{data}{Data frame or a numeric vector.}

\item{method}{\code{"pearson"} (default), \code{"spearman"}, or \code{"kendall"}.}

\item{nsmall}{Number of decimal places of output. Default is \code{2}.}

\item{p.adjust}{Adjustment of \emph{p} values for multiple tests: \code{"none", "fdr", "holm", "bonferroni", ...}
For details, see \code{stats::\link[stats]{p.adjust}}.}

\item{plot}{\code{TRUE} (default) or \code{FALSE}, plot the correlation matrix.}

\item{plot.range}{Range of correlation coefficients for plot. Default is \code{c(-1, 1)}.}

\item{plot.color}{Color gradient for plot. Default is \code{c("#B52127", "white", "#2171B5")}.
You may also set it to, e.g., \code{c("red", "white", "blue")}.}

\item{save.file}{\code{NULL} (default, plot in RStudio) or a file name (\code{"xxx.png"}).}

\item{width}{Width (in "inch") of the saved plot. Default is \code{8}.}

\item{height}{Height (in "inch") of the saved plot. Default is \code{6}.}

\item{dpi}{DPI (dots per inch) of the saved plot. Default is \code{500}.}
}
\description{
Correlation analysis with test and plot.
}
\examples{
Corr(airquality)
Corr(airquality, p.adjust="bonferroni")
# Corr(airquality, save.file="Air-Corr.png")

Corr(bfi[c("gender", "age", "education")])
# Corr(bfi, save.file="BFI-Corr.png", width=9, height=9)

}
