% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{Corr}
\alias{Corr}
\title{Correlation analysis with test and plot}
\usage{
Corr(data, method = "pearson", digits = 4, short = FALSE,
  plot = TRUE)
}
\description{
Correlation analysis with test and plot
}
