% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{Describe}
\alias{Describe}
\title{Descriptive statistics}
\usage{
Describe(data, nsmall = 2, plot = FALSE, smooth = FALSE,
  save.file = NULL, save.size = "8:6", save.dpi = 500)
}
\arguments{
\item{data}{\code{data.frame} or \code{data.table}.}

\item{nsmall}{Number of decimal places of output. Default is 2.}

\item{plot}{\code{TRUE} or \code{FALSE} (default), visualize the descriptive statistics with \code{GGally::\link[GGally]{ggpairs}}.}

\item{smooth}{\code{FALSE} (default), \code{"lm"}, or \code{"loess"}, add fitted lines to scatter plots (if any).}

\item{save.file}{\code{NULL} (default, plot in RStudio) or a file name (\code{"xxx.png"}).}

\item{save.size}{Size (in "inch") of the saved plot. Default is \code{"8:6"}.}

\item{save.dpi}{DPI (dots per inch) of the saved plot. Default is \code{500}.}
}
\description{
Descriptive statistics
}
\examples{
d=bfi
d$gender=as.factor(d$gender)
d$education=as.factor(d$education)
Describe(d[c("age", "gender", "education")])
Describe(d[c("age", "gender", "education")], plot=T)

Describe(airquality, plot=T, smooth="lm",
         save.file="Desc.png", save.size="10:8", save.dpi=1000)
}
