% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_plot.R
\name{drawChinaMap}
\alias{drawChinaMap}
\title{Draw standard China maps}
\usage{
drawChinaMap(provdata = NULL, citydata = NULL, var = NA,
  multiply = 1, log = FALSE, nsmall = 0, colors = "Blues",
  direc = 1, addlabel = TRUE, labelprefix = "", labelseg = ":",
  tag = "", title = var, guidetitle = "", addguidevalue = TRUE,
  limits = NULL, breaks = NULL, bordersize = 0.2,
  bordercolor = "grey70", na.color = "grey90",
  filename = "ChinaMap.png", dpi = 500)
}
\arguments{
\item{provdata}{Province-level data. You can use \code{\link[dplyr]{left_join}} to merge your prov data with \code{provdata_demo} (a demo dataset in \code{bruceR}) by the variable \code{"prov"}.
If not specified, it will draw a demo map for you (see Examples).

For details about \code{provdata_demo}, type this in your console:
\code{View(provdata_demo)}}

\item{citydata}{City-level data with two variables (must be "geoE" and "geoN") specifying the longitude and latitude of cities, respectively.}

\item{var}{A character specifying the variable you want to map to the plot.}

\item{multiply}{A number useful when you want to expand the raw values by, e.g., 100 times.}

\item{log}{\code{TRUE} or \code{FALSE} (default). Whether to log-transform the raw values.}

\item{nsmall}{Number of decimal places of output. Default is 0.}

\item{colors}{Color palettes. The following palettes are available (see \code{\link[ggplot2]{scale_color_brewer}}):

\strong{Sequential:}
\code{Blues, Greens, Greys, Oranges, Purples, Reds,
BuGn, BuPu, GnBu, OrRd, PuBu, PuRd, RdPu, YlGn,
PuBuGn, YlGnBu, YlOrBr, YlOrRd}

\strong{Diverging:}
\code{BrBG, PiYG, PRGn, PuOr, RdBu, RdGy, RdYlBu, RdYlGn, Spectral}

\strong{Qualitative (not suggested):}
\code{Accent, Dark2, Paired, Pastel1, Pastel2, Set1, Set2, Set3}}

\item{direc}{\code{1} (default) or \code{-1}, specifying the direction of color palette.}

\item{addlabel}{\code{TRUE} (default) or \code{FALSE}. Whether to add value labels. For clarity, value labels are only added to provinces but not to cities.}

\item{labelprefix}{A character specifying a variable for adding label prefix, usually \code{"prov"} if you want to add the names of provinces prior to values.}

\item{labelseg}{A character specifying the joint character between label prefix and values (e.g., setting to \code{": "} will make a label look like \code{"Beijing: 1.23"}).}

\item{tag}{Tag of the map (left-top corner).}

\item{title}{Title of the map.}

\item{guidetitle}{Title of the colorbar guide.}

\item{addguidevalue}{\code{TRUE} (default) or \code{FALSE}. Whether to add values under the colorbar guide.}

\item{limits}{A number vector specifying the range of values to plot (relevant both to the main plot and to the colorbar guide). Default is the actual range of your variable.}

\item{breaks}{A number vector specifying the breaking points of colorbar, e.g., \code{seq(0, 100, 25)}.}

\item{bordersize}{Line size of map border. Default is \code{0.2}.}

\item{bordercolor}{Line color of map border. Default is \code{"grey70"}.}

\item{na.color}{A color for those provinces with missing values. Default is \code{"grey90"}.}

\item{filename}{File name to create on disk. The file type can be any of ".pdf", ".png", ".jpg", ".bmp", ".tiff", ".eps", ... (see \code{\link[ggplot2]{ggsave}}).}

\item{dpi}{Dots per inch (DPI). A higher DPI produces clearer and more detailed output. Academic papars usually require 300 dpi at least. Here I use 500 as a default value.
(Note: PDF documents are not influenced by DPI.)}
}
\value{
Invisibly return a list of two maps (a main map and a sub-map for Nanhai islands).
}
\description{
Draw standard China maps
}
\examples{
drawChinaMap() # draw a demo map
drawChinaMap(provdata_demo, var="geoE", nsmall=1, filename="ChinaMap1.png")
drawChinaMap(provdata_demo, var="geoN", nsmall=1, colors="Reds", direc=-1, addlabel=FALSE, filename="ChinaMap2.png")
}
