% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{dtime}
\alias{dtime}
\title{Timer (compute time difference)}
\usage{
dtime(t0, unit = "secs")
}
\arguments{
\item{t0}{Time at the beginning.}

\item{unit}{A value from \code{c("auto", "secs", "mins", "hours", "days", "weeks")}. Default is \code{"secs"}.}
}
\description{
Timer (compute time difference)
}
\examples{
t0=Sys.time()
dtime(t0)
}
