% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{dtime}
\alias{dtime}
\title{Timer}
\usage{
dtime(t0, unit = "secs")
}
\arguments{
\item{t0}{Time at the beginning, typically: \code{t0 = Sys.time()}}

\item{unit}{A value from \code{c("auto", "secs", "mins", "hours", "days", "weeks")}. Default is \code{"secs"}.}
}
\value{
A string.
}
\description{
Timer
}
\examples{
t0=Sys.time()
dtime(t0)
}
