% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{dtime}
\alias{dtime}
\title{Timer (compute time difference).}
\usage{
dtime(t0, unit = "secs", nsmall = 0)
}
\arguments{
\item{t0}{Time at the beginning.}

\item{unit}{Options: \code{"auto", "secs", "mins", "hours", "days", "weeks"}. Default is \code{"secs"}.}

\item{nsmall}{Number of decimal places of output. Default is \code{0}.}
}
\value{
A character string of time difference.
}
\description{
Timer (compute time difference).
}
\examples{
t0=Sys.time()
dtime(t0)

}
