% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_manova.R
\name{EMMEANS}
\alias{EMMEANS}
\title{Simple-effect analysis (for interaction) and post-hoc pairwise comparison (for factors with 3 or more levels)}
\usage{
EMMEANS(model, effect = NULL, by = NULL, p.adjust = "bonferroni")
}
\arguments{
\item{p.adjust}{Adjustment of \emph{p} values for multiple comparisons. Default is \code{"bonferroni"}.
You may set it to \code{"none", "fdr", "hochberg", "hommel", "holm", "tukey", "bonferroni"}.
For details, see \code{\link[stats]{p.adjust}}.}
}
\description{
Simple-effect analysis (for interaction) and post-hoc pairwise comparison (for factors with 3 or more levels)
}
\examples{
## Between-Subject Design

View(between.1)
MANOVA(data=between.1, dv="SCORE", between="A") \%>\%
  EMMEANS("A")

View(between.2)
MANOVA(data=between.2, dv="SCORE", between=c("A", "B")) \%>\%
  EMMEANS("A", by="B") \%>\%
  EMMEANS("B", by="A")

View(between.3)
MANOVA(data=between.3, dv="SCORE", between=c("A", "B", "C")) \%>\%
  EMMEANS(c("A", "B"), by="C") \%>\%
  EMMEANS("A", by=c("B", "C"))


## Within-Subject Design

View(within.1)
MANOVA(data=within.1, dvs="A1:A4", dvs.pattern="A(.)",
       within="A") \%>\%
  EMMEANS("A")

View(within.2)
MANOVA(data=within.2, dvs="A1B1:A2B3", dvs.pattern="A(.)B(.)",
       within=c("A", "B")) \%>\%
  EMMEANS("A", by="B") \%>\%
  EMMEANS("B", by="A")

View(within.3)
MANOVA(data=within.3, dvs="A1B1C1:A2B2C2", dvs.pattern="A(.)B(.)C(.)",
       within=c("A", "B", "C")) \%>\%
  EMMEANS(c("A", "B"), by="C") \%>\%
  EMMEANS("A", by=c("B", "C"))


## Mixed Design

View(mixed.2_1b1w)
MANOVA(data=mixed.2_1b1w, dvs="B1:B3", dvs.pattern="B(.)",
       between="A", within="B", correction="GG") \%>\%
  EMMEANS("A", by="B") \%>\%
  EMMEANS("B", by="A")
}
\seealso{
\code{\link{MANOVA}}
}
