% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_utils.R
\name{formatF}
\alias{formatF}
\title{Format numeric values.}
\usage{
formatF(x, digits = 3)
}
\arguments{
\item{x}{A number or numeric vector.}

\item{digits}{Number of decimal places of output. Defaults to \code{3}.}
}
\value{
Formatted character string.
}
\description{
Format numeric values.
}
\examples{
formatF(pi, 20)

}
\seealso{
\code{\link[base:format]{format}}, \code{\link{formatN}}
}
