% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{formatF}
\alias{formatF}
\title{Format numeric values}
\usage{
formatF(x, nsmall = 3)
}
\arguments{
\item{x}{any R object, typically numeric.}
}
\value{
Formatted R object.
}
\description{
Format numeric values
}
