% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{formatF}
\alias{formatF}
\title{Format numeric values}
\usage{
formatF(x, nsmall = 3)
}
\arguments{
\item{x}{Any R object, typically numeric.}

\item{nsmall}{Number of decimal places of output. Default is 3.}
}
\value{
Formatted character string.
}
\description{
Format numeric values
}
\examples{
formatF(pi)
}
