% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{formatF}
\alias{formatF}
\title{Format numeric values.}
\usage{
formatF(x, nsmall = 3)
}
\arguments{
\item{x}{A number or numeric vector.}

\item{nsmall}{Number of decimal places of output. Default is \code{3}.}
}
\value{
Formatted character string.
}
\description{
Format numeric values.
}
\examples{
formatF(pi, 20)

}
\seealso{
\code{\link{formatN}}
}
