% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{formatN}
\alias{formatN}
\title{Format "1234" to "1,234"}
\usage{
formatN(x, mark = ",")
}
\arguments{
\item{x}{any R object, typically numeric.}
}
\value{
Formatted R object.
}
\description{
Format "1234" to "1,234"
}
