% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR-stats_4_regress.R
\name{formula_expand}
\alias{formula_expand}
\title{Expand all interaction terms in a formula.}
\usage{
formula_expand(formula, as.char = FALSE)
}
\arguments{
\item{formula}{R formula or a character string indicating the formula.}

\item{as.char}{Return character? Default is \code{FALSE}.}
}
\value{
A formula/character object including all expanded terms.
}
\description{
Expand all interaction terms in a formula.
}
\examples{
formula_expand(y ~ a*b*c)
formula_expand("y ~ a*b*c")

}
