% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_regress.R
\name{formula_expand}
\alias{formula_expand}
\title{Expand all interaction terms in a formula (of \code{lm, glm, lmer, glmer})}
\usage{
formula_expand(formula)
}
\description{
Expand all interaction terms in a formula (of \code{lm, glm, lmer, glmer})
}
\examples{
formula_expand(y ~ a*b*c)
}
