% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{Freq}
\alias{Freq}
\title{Frequency statistics with histogram and density plot}
\usage{
Freq(var, label = NULL, sort = "", nsmall = 1, plot = FALSE,
  bins = 18, fill.color = "#FDF6E3")
}
\description{
Frequency statistics with histogram and density plot
}
\examples{
Freq(psych::bfi$education)
Freq(psych::bfi$gender, label=c("Male", "Female"))
Freq(psych::bfi$age, plot=T)
}
