% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_stats_01_basic.R
\name{Freq}
\alias{Freq}
\title{Frequency statistics.}
\usage{
Freq(var, label = NULL, sort = "", digits = 1, nsmall = digits, file = NULL)
}
\arguments{
\item{var}{Variable (or a vector of values).}

\item{label}{[Optional] A vector re-defining the labels of values.}

\item{sort}{\code{""} (default, sorted by the order of variable values/labels), \code{"-"} (decreasing by N), or \code{"+"} (increasing by N).}

\item{digits, nsmall}{Number of decimal places of output. Default is \code{1}.}

\item{file}{File name of MS Word (\code{.doc}).}
}
\value{
A data frame of frequency statistics.
}
\description{
Frequency statistics.
}
\examples{
data=psych::bfi
Freq(data$education)
Freq(data$gender, label=c("Male", "Female"))
Freq(data$age)

}
