% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_stats_01_basic.R
\name{Freq}
\alias{Freq}
\title{Frequency statistics (to R Console or MS Word).}
\usage{
Freq(var, label = NULL, sort = "", nsmall = 1, file = NULL)
}
\arguments{
\item{var}{Vector or variable.}

\item{label}{[optional] A vector re-defining the labels of values.}

\item{sort}{\code{""} (default, sorted by raw order), \code{"-"} (decreasing), or \code{"+"} (increasing).}

\item{nsmall}{Number of decimal places of output. Default is \code{1}.}

\item{file}{File name of MS Word (\code{.doc}).}
}
\value{
A data frame of frequency statistics.
}
\description{
Frequency statistics (to R Console or MS Word).
}
\examples{
Freq(bfi$education)
Freq(bfi$gender, label=c("Male", "Female"))
Freq(bfi$age)

}
