% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_utils.R
\name{\%partin\%}
\alias{\%partin\%}
\title{A simple extension of \code{\%in\%}.}
\usage{
pattern \%partin\% vector
}
\arguments{
\item{pattern}{Character string containing \strong{regular expressions} to be matched.}

\item{vector}{Character vector.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
A simple extension of \code{\%in\%}.
}
\examples{
"Bei" \%partin\% c("Beijing", "Shanghai")  # TRUE
"bei" \%partin\% c("Beijing", "Shanghai")  # FALSE
"[aeiou]ng" \%partin\% c("Beijing", "Shanghai")  # TRUE

}
\seealso{
\code{\link[base:match]{\%in\%}},
\code{\link{\%allin\%}},
\code{\link{\%anyin\%}},
\code{\link{\%nonein\%}}
}
