% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_regress.R
\name{group_mean_center}
\alias{group_mean_center}
\title{Group-mean centering}
\usage{
group_mean_center(data, vars, by)
}
\arguments{
\item{data}{a \code{data.frame} or \code{data.table}}

\item{vars}{a character vector specifying which variable(s) to be centered}

\item{by}{a character indicating the grouping variable}
}
\value{
A \code{data.frame} or \code{data.table}, with the same type as the input data.
}
\description{
Compute group-mean centered variables. Usually used for HLM level-1 variables.
}
\seealso{
\code{\link{grand_mean_center}}
}
