% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{LOOKUP}
\alias{LOOKUP}
\title{Search, match, and look up values (like Excel's functions \code{INDEX + MATCH})}
\usage{
LOOKUP(data, vars, data.ref, vars.ref, vars.lookup)
}
\arguments{
\item{data}{\code{data.frame} or \code{data.table}.}

\item{vars}{Character or character vector, specifying the variable(s) to be searched in \code{data}.}

\item{data.ref}{Reference data containing both the reference variable(s) and the lookup variable(s).}

\item{vars.ref}{Character or character vector (with the \strong{same length and order} as \code{vars}),
specifying the reference variable(s) to be matched in \code{data.ref}.}

\item{vars.lookup}{Character or character vector, specifying the variable(s) to be looked up and returned from \code{data.ref}.}
}
\value{
A \code{data.frame} or \code{data.table} with the lookup values added.
If multiple values were simultaneously matched, a warning message would be printed.
}
\description{
In Excel, we can use \code{VLOOKUP}, \code{HLOOKUP}, \code{XLOOKUP} (a new function released in 2019), or the combination of \code{INDEX} and \code{MATCH} to search, match, and look up values.
Here I provide a similar function.
}
\examples{
dict=data.table(City=rep(c("A", "B", "C"), each=5),
                Year=rep(2013:2017, times=3),
                GDP=RANDBETWEEN(1000:2000, n=15, seed=1),
                PM2.5=RANDBETWEEN(10:300, n=15, seed=1))
dict

data=data.table(sub=1:5,
                city=c("A", "A", "B", "C", "C"),
                year=c(2013, 2014, 2015, 2016, 2017))
data

# LOOKUP(data, "city", dict, "City", "GDP")  # return with a warning
LOOKUP(data, c("city", "year"), dict, c("City", "Year"), "GDP")
LOOKUP(data, c("city", "year"), dict, c("City", "Year"), c("GDP", "PM2.5"))
}
\seealso{
\href{https://support.office.com/en-us/article/XLOOKUP-function-B7FD680E-6D10-43E6-84F9-88EAE8BF5929}{XLOOKUP: Excel's new function, the VLOOKUP "slayer" (August 2019)}
\href{https://www.excel-university.com/xlookup/}{XLOOKUP: Excel University}
}
