% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_manova.R
\name{MANOVA}
\alias{MANOVA}
\title{Multivariate ANOVA}
\usage{
MANOVA(data, dv = NULL, dvs = NULL, dvs.pattern = "",
  between = NULL, within = NULL, covariate = NULL,
  which.observed = NULL, sph.correction = "none", subID = NULL)
}
\arguments{
\item{data}{Any data object (e.g., \code{data.frame, data.table}).
\itemize{
  \item You can input a \strong{wide-format} data (i.e., one subject occupies one row, and repeated measures occupy multiple columns).
  Then, if your data have repeated measures, the function can \strong{\emph{automatically}} transform it into a \strong{long-format} data.
  \item If you input a \strong{long-format} data, please also specify \strong{subID} (i.e., the variable indicating different subjects).
}}

\item{dv}{Variable name of dependent variable in your data.
\itemize{
  \item If you input a \strong{wide-format} data, then \code{dv} is only for between-subjects design.
  For designs with repeated measures, please use \code{dvs} and \code{dvs.pattern} instead.
  \item If you input a \strong{long-format} data, then it is just the outcome variable.
}}

\item{dvs}{\strong{[only for "wide-format" data and designs with repeated measures]}

Variable names of repeated measures in your data.
\itemize{
  \item You can use \code{":"} to specify a range of vairables: e.g., \code{"A1B1:A2B3"}
  (similar to the SPSS syntax "TO"; the variables should be put in order)
  \item You can also use a character vector to specify variable names: e.g., \code{c("Cond1", "Cond2", "Cond3")}
}}

\item{dvs.pattern}{\strong{[only for "wide-format" data and designs with repeated measures]}

If you set \code{dvs}, you must also set the pattern of variable names by using \href{https://www.jb51.net/shouce/jquery1.82/regexp.html}{regular expressions}.

\strong{Examples:}
\itemize{
  \item \code{"Cond(.)"} can extract levels from \code{"Cond1", "Cond2", "Cond3", ...}

  \strong{You can rename the factor name} by using \code{within}: e.g., \code{within="Condition"}
  \item \code{"X(..)Y(..)"} can extract levels from \code{"X01Y01", "X02Y02", "XaaYbc", ...}
  \item \code{"X(.+)Y(.+)"} can extract levels from \code{"X1Y1", "XaYb", "XaY002", ...}
}

\strong{Tips on regular expression:}
\itemize{
  \item \code{"(.)"} extracts any single character (can be number, letter, or other symbols)
  \item \code{"(.+)"} extracts >= 1 character(s)
  \item \code{"(.*)"} extracts >= 0 character(s)
  \item \code{"([0-9])"} extracts any single number
  \item \code{"([a-z])"} extracts any single letter
  \item each pair of \code{"()"} extracts levels for each factor
}}

\item{between}{Between-subjects factors. Character string (e.g., \code{"A"}) or vector (e.g., \code{c("A", "B")}). Default is \code{NULL}.}

\item{within}{Within-subjects factors. Character string (e.g., \code{"A"}) or vector (e.g., \code{c("A", "B")}). Default is \code{NULL}.}

\item{covariate}{Covariates (if necessary). Character string or vector. Default is \code{NULL}.}

\item{which.observed}{\strong{[only effective for computing generalized \eqn{\eta^2}]}

Factors that are observed or measured (e.g., gender, age group, measured covariates) but not experimentally manipulated. Default is \code{NULL}.
The generalized \eqn{\eta^2} requires correct specification of the observed (vs. manipulated) variables.
(If all the variables in \code{between} and \code{within} are set to \code{observed}, then generalized \eqn{\eta^2} will be equal to \eqn{\eta^2}.)}

\item{sph.correction}{\strong{[only effective for repeated measures with >= 3 levels]}

Sphericity correction method to adjust the degrees of freedom (\emph{df}) when the sphericity assumption is violated. Default is \code{"none"}.
If Mauchly's test of sphericity is significant, you may set it to \code{"GG"} (Greenhouse-Geisser) or \code{"HF"} (Huynh-Feldt).}

\item{subID}{As mentioned above, if your input is already a \strong{long-format} data, then you should also specify the subject ID because one subject occupies multiple rows.
However, if you input a \strong{wide-format} data, you \strong{don't need} to do that.}
}
\description{
Easily perform MANOVA (between-subjects, within-subjects, and mixed design).
}
\details{
This function is based on and extends the \code{\link[afex]{aov_ez}} function in the R package \code{afex}.
You only need to specify the data, dependent variable(s), and factors (between-subjects and/or within-subjects).
Then, almost all the outputs you need will be displayed in an elegant manner, including effect sizes (partial \eqn{\eta^2}) and their confidence intervals (CIs).
90\% CIs for partial \eqn{\eta^2} are reported, following the suggestion by Steiger (2004).

In addition to partial \eqn{\eta^2}, it will also output many other effect-size measures:
\eqn{\eta^2}, generalized \eqn{\eta^2}, \eqn{\omega^2}, and Cohen's \emph{f}.
For statistical details, see \url{https://en.wikipedia.org/wiki/Effect_size}

\strong{Demo Datasets:}

The demo datasets were obtained from a course of "multifactor experimental design" in \emph{Beijing Normal University} (2016).
In this course, we used a book written by Prof. Hua Shu (\strong{\emph{"Factorial Experimental Design in Psychological and Educational Research"}}).
The book provided a seires of demo datasets to show different experimental designs and how to do MANOVA in SPSS.
Here, we reuse these excellent demo datasets to show how the R function \code{MANOVA} can easily handle almost all types of designs.
Many thanks go to the contributors of the datasets.
\describe{
  \item{\strong{1. Between-Subjects Design}}{
    \itemize{
      \item \code{between.1} - A(4)
      \item \code{between.2} - A(2) * B(3)
      \item \code{between.3} - A(2) * B(2) * C(2)
    }
  }
  \item{\strong{2. Within-Subjects Design}}{
    \itemize{
      \item \code{within.1} - A(4)
      \item \code{within.2} - A(2) * B(3)
      \item \code{within.3} - A(2) * B(2) * C(2)
    }
  }
  \item{\strong{3. Mixed Design}}{
    \itemize{
      \item \code{mixed.2_1b1w} - A(2, between) * B(3, within)
      \item \code{mixed.3_1b2w} - A(2, between) * B(2, within) * C(2, within)
      \item \code{mixed.3_2b1w} - A(2, between) * B(2, within) * C(2, between)
    }
  }
}
}
\examples{
#### Between-Subjects Design ####

View(between.1)
MANOVA(data=between.1, dv="SCORE", between="A")

View(between.2)
MANOVA(data=between.2, dv="SCORE", between=c("A", "B"))

View(between.3)
MANOVA(data=between.3, dv="SCORE", between=c("A", "B", "C"))


#### Within-Subjects Design ####

View(within.1)
MANOVA(data=within.1, dvs="A1:A4", dvs.pattern="A(.)",
       within="A")
## the same:
MANOVA(data=within.1, dvs=c("A1", "A2", "A3", "A4"), dvs.pattern="A(.)",
       within="MyFactor")  # renamed the within-subjects factor

View(within.2)
MANOVA(data=within.2, dvs="A1B1:A2B3", dvs.pattern="A(.)B(.)",
       within=c("A", "B"))

View(within.3)
MANOVA(data=within.3, dvs="A1B1C1:A2B2C2", dvs.pattern="A(.)B(.)C(.)",
       within=c("A", "B", "C"))


#### Mixed Design ####

View(mixed.2_1b1w)
MANOVA(data=mixed.2_1b1w, dvs="B1:B3", dvs.pattern="B(.)",
       between="A", within="B")
MANOVA(data=mixed.2_1b1w, dvs="B1:B3", dvs.pattern="B(.)",
       between="A", within="B", sph.correction="GG")

View(mixed.3_1b2w)
MANOVA(data=mixed.3_1b2w, dvs="B1C1:B2C2", dvs.pattern="B(.)C(.)",
       between="A", within=c("B", "C"))

View(mixed.3_2b1w)
MANOVA(data=mixed.3_2b1w, dvs="B1:B2", dvs.pattern="B(.)",
       between=c("A", "C"), within="B")


#### Other Examples ####
data.new=mixed.3_1b2w
names(data.new)=c("Group", "Cond_01", "Cond_02", "Cond_03", "Cond_04")
MANOVA(data=data.new, dvs="Cond_01:Cond_04", dvs.pattern="Cond_(..)",
       between="Group", within="Condition")  # renamed the within-subjects factor
}
\references{
Olejnik, S., & Algina, J. (2003). Generalized eta and omega squared statistics: Measures of effect size for some common research designs.
\emph{Psychological Methods, 8}(4), 434-447. \url{https://doi.org/10.1037/1082-989X.8.4.434}

Steiger, J. H. (2004). Beyond the F test: Effect size confidence intervals and tests of close fit in the analysis of variance and contrast analysis.
\emph{Psychological Methods, 9}(2), 164-182. \url{https://doi.org/10.1037/1082-989X.9.2.164}
}
\seealso{
\code{\link{EMMEANS}}
}
