% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{pkg_depend}
\alias{pkg_depend}
\title{Check package dependencies}
\usage{
pkg_depend(pkg)
}
\arguments{
\item{pkg}{package name}
}
\description{
Check package dependencies
}
\examples{
pkg.depend("sjstats")
pkg.depend("MuMIn")
}
