% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{pkg_depend}
\alias{pkg_depend}
\title{Check dependencies of a package.}
\usage{
pkg_depend(pkg, exclude = NULL)
}
\arguments{
\item{pkg}{Package name (a character string).}

\item{exclude}{[optional] Package(s) and their dependencies excluded from the dependencies of \code{pkg}.
Useful if you want to see the unique contribution of \code{pkg}.}
}
\description{
Check dependencies of a package.
}
\examples{
## Not run:

# pkg_depend("jmv")
# pkg_depend("dplyr", exclude="jmv")  # no unique contribution

## End(Not run)

}
