% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{pkg_depend}
\alias{pkg_depend}
\title{Check dependencies of R packages.}
\usage{
pkg_depend(pkgs, excludes = NULL)
}
\arguments{
\item{pkgs}{Package(s).}

\item{excludes}{[optional] Package(s) and their dependencies excluded from the dependencies of \code{pkgs}.
Useful if you want to see the unique dependencies of \code{pkgs}.}
}
\value{
A character vector of package names.
}
\description{
Check dependencies of R packages.
}
\seealso{
\code{\link{pkg_install_suggested}}
}
