% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{pkg_install_suggested}
\alias{pkg_install_suggested}
\title{Install suggested R packages.}
\usage{
pkg_install_suggested(by = "bruceR")
}
\arguments{
\item{by}{Suggested by which package? Default is \code{"bruceR"}.}
}
\value{
No return value.
}
\description{
It checks and installs R packages suggested by \code{bruceR} (default)
or any other package.
}
\seealso{
\link{pkg_depend}
}
