% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_utils.R
\name{pkg_install_suggested}
\alias{pkg_install_suggested}
\title{Install suggested R packages.}
\usage{
pkg_install_suggested(by)
}
\arguments{
\item{by}{Suggested by which package?}
}
\value{
No return value.
}
\description{
Install suggested R packages.
}
\examples{
\dontrun{

pkg_install_suggested()  # install all packages suggested by me
}

}
\seealso{
\code{\link{pkg_depend}}
}
