% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{Print}
\alias{Print}
\alias{Glue}
\title{Print a string with fruitful formats in a concise manner}
\usage{
Print(...)

Glue(...)
}
\arguments{
\item{...}{A string in which expressions enclosed by \code{"{ }"} will be evaluated as R code.
Long strings are broken by line and concatenated together. Leading whitespace and blank lines from the first and last lines are automatically trimmed.}
}
\value{
A string.
}
\description{
Be tired of \code{print()} and \code{cat()}? Try this function! Just type \code{example("Print")} in the console and see its power.
}
\details{
This function is based on \code{\link[glue]{glue}} and \code{\link[glue]{glue_col}}.
See more details in their help pages.
}
\section{Functions}{
\itemize{
\item \code{Glue}: Glue strings
}}

\examples{
name="Bruce"
Print("My name is <<underline <<bold {name}>>>>.
       <<bold <<blue Pi = {pi:.15}.>>>>
       <<italic <<green 1 + 1 = {1 + 1}.>>>>
       sqrt({x}) = <<red {sqrt(x):.3}>>", x=10)
}
