% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{print_table}
\alias{print_table}
\title{Print three-line table}
\usage{
print_table(x, row.names = TRUE, nsmalls = 3, style = c("line",
  "data"))
}
\arguments{
\item{x}{a matrix, data.frame, data.table, or a model (\code{lm, glm, lmer, glmer, ...}).}
}
\description{
Print three-line table
}
\examples{
model=lm(Temp ~ Month + Day + Wind + Solar.R, data=airquality)
print_table(model)
}
