% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{RANDBETWEEN}
\alias{RANDBETWEEN}
\title{Random sampling (like Excel's function \code{RANDBETWEEN}).}
\usage{
RANDBETWEEN(range, n = 1, seed = NULL)
}
\arguments{
\item{range}{Numeric or character vector.}

\item{n}{Sample size for sampling. Default is \code{1}.}

\item{seed}{Random seed.}
}
\value{
Numeric or character vector (the same class as \code{range}).
}
\description{
Random sampling (like Excel's function \code{RANDBETWEEN}).
}
\examples{
RANDBETWEEN(1:10, n=1000) \%>\% Freq()
RANDBETWEEN(LETTERS, n=1000) \%>\% Freq()

}
