% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{Run}
\alias{Run}
\title{Run code parsed from text.}
\usage{
Run(..., silent = FALSE)
}
\arguments{
\item{...}{Character string(s) to run.
You can use \code{"{ }"} to insert any R object in the environment.}

\item{silent}{Suppress error/warning messages. Default is \code{FALSE}.}
}
\value{
Invisibly return the running expression(s).
}
\description{
Run code parsed from text.
}
\examples{
Run("a=1", "b=2")
Run("print({a+b})")

}
