% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{set.wd}
\alias{set.wd}
\title{Set working directory to the path of \strong{current} script}
\usage{
set.wd(dir = NULL)
}
\arguments{
\item{dir}{A character string. If \code{NULL} (default), set working directory to the path of the current R script.}
}
\description{
Set working directory to the path of \strong{current} script
}
\examples{
set.wd()  # set working directory to the path of the current R script
set.wd("E:/")  # "\\" is not allowed, you should use "/"
set.wd("../")  # set working directory to the parent directory
}
