% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_basic.R
\name{set.wd}
\alias{set.wd}
\title{Set working directory to where the current R script is.}
\usage{
set.wd(dir = NULL)
}
\arguments{
\item{dir}{\code{NULL} (default) or a character string specifying the working directory.
If \code{NULL}, set working directory to the path of \strong{the current R script}.}
}
\value{
No return value.
}
\description{
Set working directory to where the current R script is.
}
\examples{
\dontrun{

set.wd()  # set working directory to the path of the current R script
set.wd("D:/")  # or any other path (use "/" rather than "\")
set.wd("../")  # set working directory to the parent directory

}

}
\seealso{
\code{\link{setwd}}
}
