% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruceR_plot.R
\name{theme_bruce}
\alias{theme_bruce}
\title{A nice \code{ggplot2} theme for scientific plotting}
\usage{
theme_bruce(base.size = 12, line.size = 0.5, border = "",
  bg = "white", tag = "bold", plot.title = "bold",
  axis.title = "plain", title.pos = 0, caption.pos = 1,
  grid.x = "", grid.y = "grey90", line.x = TRUE, line.y = TRUE,
  tick.x = TRUE, tick.y = TRUE)
}
\arguments{
\item{base.size}{Basic font size. Default is 12.}

\item{line.size}{Line width. Default is 0.5.}

\item{border}{\code{F}, \code{""} (default), or a color. If set to \code{F} or \code{""}, it will not draw the border.}
}
\description{
A nice \code{ggplot2} theme for scientific plotting
}
